Update: split file env_expansion
